Crate json_rpc_stdio

This crate implements JSON-RPC over standard IO. It uses tokio for async IO, and jsonrpc_core for the JSON-RPC part.

Notifications in the Client are not supported yet.

Structs

Client

A handle to a connected client you can use to send requests.

ClientAdapter

The other side of the channels. Only used as a handle to be passed into run_with_client().

Functions

new_client

Constructor a JSON-RPC client. Returns a tuple: the client you can use to send requests, and the adapter you must pass to run_with_client() to connect the client to the proper IO.

run

Start doing JSON-RPC over stdio, server-only. The future will only return once stdin is closed or another IO error happens.

run_with_client

Start doing JSON-RPC over stdio. The future will only return once stdin is closed or another IO error happens.